updating pull request as per review
