+++ webv2/common/utilities.go
@@ -25,6 +25,7 @@ import (
 
 	database "cloud.google.com/go/spanner/admin/database/apiv1"
 	"github.com/cloudspannerecosystem/harbourbridge/conversion"
+	"github.com/cloudspannerecosystem/harbourbridge/internal"
 	adminpb "google.golang.org/genproto/googleapis/spanner/admin/database/v1"
 )
 
@@ -111,17 +112,6 @@ func createDatabase(ctx context.Context, uri string) error {
 	return nil
 }
 
-// Contain check string is present in given list.
-func Contain(fc []string, col string) string {
-
-	for _, s := range fc {
-		if s == col {
-			return col
-		}
-	}
-	return ""
-}
-
 // DuplicateInArray checks if there is any duplicate element present in the list.
 func DuplicateInArray(element []int) int {
 	visited := make(map[int]bool, 0)
@@ -154,4 +144,65 @@ func Difference(listone, listtwo []int) []int {
 		}
 	}
 	return diff
+}
+
+// IsColumnPresent check string is present in given list.
+func IsColumnPresent(columns []string, col string) string {
+
+	for _, c := range columns {
+		if c == col {
+			return col
+		}
+	}
+	return ""
+}
+
+// RemoveSchemaIssue removes issue from the given list.
+func RemoveSchemaIssue(schemaissue []internal.SchemaIssue, issue internal.SchemaIssue) []internal.SchemaIssue {
+
+	for i := 0; i < len(schemaissue); i++ {
+		if schemaissue[i] == issue {
+			schemaissue = append(schemaissue[:i], schemaissue[i+1:]...)
+		}
+	}
+	return schemaissue
+}
+
+// IsSchemaIssuePresent checks if issue is present in the given schemaissue list.
+func IsSchemaIssuePresent(schemaissue []internal.SchemaIssue, issue internal.SchemaIssue) bool {
+
+	for _, s := range schemaissue {
+		if s == issue {
+			return true
+		}
+	}
+	return false
+}
+
+// RemoveSchemaIssues remove all  hotspot and interleaved from given list.
+// RemoveSchemaIssues is used when we are adding or removing primary key column from primary key.
+func RemoveSchemaIssues(schemaissue []internal.SchemaIssue) []internal.SchemaIssue {
+
+	switch {
+
+	case IsSchemaIssuePresent(schemaissue, internal.HotspotAutoIncrement):
+		schemaissue = RemoveSchemaIssue(schemaissue, internal.HotspotAutoIncrement)
+		fallthrough
+
+	case IsSchemaIssuePresent(schemaissue, internal.HotspotTimestamp):
+		schemaissue = RemoveSchemaIssue(schemaissue, internal.HotspotTimestamp)
+		fallthrough
+
+	case IsSchemaIssuePresent(schemaissue, internal.InterleavedOrder):
+		schemaissue = RemoveSchemaIssue(schemaissue, internal.InterleavedOrder)
+
+	case IsSchemaIssuePresent(schemaissue, internal.InterleavedNotInOrder):
+		schemaissue = RemoveSchemaIssue(schemaissue, internal.InterleavedNotInOrder)
+		fallthrough
+
+	case IsSchemaIssuePresent(schemaissue, internal.InterleavedAddColumn):
+		schemaissue = RemoveSchemaIssue(schemaissue, internal.InterleavedAddColumn)
+	}
+
+	return schemaissue
 }
